로또 컨트롤러 내용 유저로또 컨트롤러랑 나눔
